Added DirectedRootedTree and parallel sequencing algorithms
DirectedRootedTree is a data structure that represents
a directed rooted tree. By tree an acyclic connected graph is meant.
Directed means that every edge has a direction and the structure can not be
traversed along an edge in the direction opposite to the direction of the edge.
Rooted means that there is only one vertex that has no incoming edges.
DirectedRootedTree always has to have a root. Each tree node (vertex) can
have as many children as needed.
DirectedRootedTree provides forward iterators.

Added algorithms for building parallel sequencings of the DirectedRootedTree.
Currently, only lower (S lower) and upper (S upper) parallel sequencings can be built.
These two parallel sequencings are interesting because they define the boundaries
in which the optimal parallel sequencing is located.
For now these two algorithms are sufficient for allowing user to rearrange the
schedule in the correct boundaries, but for better UX algorithm for
building optimal parallel sequencings should be provided.

For all the functionality unit tests are provided. Coverage may not be
100% full but it checks all the reasonable cases and verifies
data structure and algorithms usability.