Added TreeSerialization module
In order to be saved to a file or transferred across the network
DirectedRootedTree should have a mechanism of being written to
and read from a stream.

TreeSerialization module provides this functionality.
The format of the written tree is as follows:

<parent node id> <value data>
<parent node id> <value data>
...

where <parent node id> is a number of the node
counting from zero by depth-first search approach.
Note that for root node <parent node id> is zero
and coincides with the root node number.
<value data> is a string representation of the node`s value.

TreeSerialization module is covered by unit-tests.