Added SequencingModel module
SequencingModel implements the Qt QAbstractItemModel interface
and provides access to parallel sequencing for Qt views.
The model requires two parallel sequencings: a lower and an upper,
that define the borders in which every element in the sequencings can be moved.

SequencingModel uses the lower sequencing as an initial sequencing
and allows views to change the positions of the elements
ensuring that every element is not moved outside its borders.
New elements addition as well as creating new columns is forbidden.

SequencingModel supports drag&drop feature that
allows a user to move items around in its borders effortlessly.
Drag&drop will be the key tool for the user
for changing SequencingModel through the view.

SequencingModelTest project has been added
for rapid devtesting of SequencingModel model.